Update VignetteIndexEntry to avoid build warning
Build warning:

Warning: Warning: The vignette title specified in \VignetteIndexEntry{} is different from the title in the YAML metadata. The former is "date_definition", and the latter is "Define the date period to download". If that is intentional, you may set options(rmarkdown.html_vignette.check_title = FALSE) to suppress this check.